pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contextflow"
version = "1.10.0"
description = "Large Language Models Tools"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Vladimir Petrukhin", email = "man4j@ya.ru" }
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
]
dependencies = [
  "transformers",
  "jinja2",
  "sseclient-py",
  "pytz"
]

[project.urls]
Homepage = "https://github.com/Equiron-AI/contextflow"
Repository = "https://github.com/Equiron-AI/contextflow"
Issues = "https://github.com/Equiron-AI/contextflow/issues"

# Указываем src-layout
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["contextflow*"]

[tool.setuptools.package-data]
contextflow = ["**/*.j2", "**/*.json", "**/*.txt"]

#[project.scripts]
#contextflow = "contextflow.__main__:main"
